requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.25.0
plotly>=5.15.0
pyyaml>=6.0
faker>=19.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
